[flang] Implement VALUE for BIND(C) calls
Fix for fir-dev issue #918. The original issue was hitting an assert
when compiling the SNAP application, because PassBy::BaseAddress-
ValueAttribute is not supported.

The PassBby::BaseAddressValueAttribute should actually be
PassBy::Value. So fix the CallInterface class to produce the
correct type. Also pass by value should not use a reference type,
so that is also fixed in the CallInterface.

Adding a simple test replicating the function that originally
caused problems.