Support array internal procedure depending on host associated symbols (#1380)
Array function results are allocated on the caller side. This requires
mapping the symbols appearing in the result specification expression on
the callee side. The current code did not account properly for the cases
where those symbols are captured symbols from the host. This lead to
"symbol is not a dummy in this call" compiler internal errors.

Handle this case and add related tests.

Also add a small TODO around SIZE intrinsic lowering for a related array function issue (it won't fire yet due to a front-end expression rewrite issue, but we will have to deal with it soon).