Support calling character functions with only the result is being explicit (#1382)
When calling a character function, the result is allocated on the caller
side. The current code was assuming that there would be an interface
when calling these kinds of function. This is not entirely correct, for
scalar character function, the result type may be explicitly defined
while the rest of the function interface is still implicit.
Support this case by relaxing an assert and avoid requiring result
symbol mapping if there is no interface (and hence no mapping to do
with a result symbol belonging to an interface).