[flang] Implement checking for DIM arguments for LBOUND and SIZE
Calls to the LBOUND and SIZE intrinsics with invalid DIM arguments were not
producing runtime errors.  Also, we were not correctly handling array arguments
for these intrinsics when the actual argument was a function call that returned
an array.

I fixed these problems by creating runtime routines for them that check the DIM
argument for validity.  I also added tests for the new runtime routines.  Also,
the changes to lowering affected the lowering code for LBOUND, UBOUND, and
SIZE.  This required updating the associated test results.