Handle dynamically optional intrinsic arguments
So far, lowering was only handling the optionality of intrinsic arguments
statically: if an optional argument was provided in the text at compile time, it
was assumed it would be present at runtime. This is not true for many intrinsic
where an optional argument as to be understood with the same semantics as
OPTIONAL argument in user functions (16.2.1 point 2). In those cases, 15.5.2.12
point 1. rules that the argument may be considered absent if it is associated
with an optional argument, a disassociated pointer, or an unallocated
allocatable. Note that none of the intrinsic arguments that are optional in that
sense have a requirement to be allocatable or pointers, which means that
15.5.2.12 point 3. (7) and (8) always apply and if the actual argument is an
optional pointer or allocatable, it cannot legally be absent, but may be
disassociated or unallocated. Some of these intrinsics are elemental.

This patch implement this in lowering (a minor commit prevents the rewrite of
CMPLX to a complex ctor in lib/evaluate when Y may be dynamically optional. This
allows keeping complex ctor branchless in lowering).

SYSTEM_CLOCK, and RANDOM_SEED, MAX and MINX are left TODOs of now. All the other
lowered intrinsics that accept such dynamic arguments are updated.