[flang] Implement the UBOUND intrinsic runtime routine
When UBOUND is called with its actual argument for ARRAY is a function call,
theappropriate runtime routine did not exist.  I implemented the processing of
the call in lowering, the runtime routine, the glue code, and a couple of
tests.

Thanks to Jean for helping to figure out how to implement this.