[optimizer] Initialize new storage in allocatable assignment
In allocatable assignment, the re-allocated storage was not initialized
before being used in the assignment. This is not correct in cases the
derived type storage must be initialized before being used (because it
has allocatable components that must be established to a valid unallocated
status, or because its components have initial values).

Instead of leaving a broken functional TODO, always call `Initialize` runtime
on derived type in allocatable assignment (it does nothing if no initialization
is required). Better type knowledge in FIR will be required to avoid
generating this calls.