allocatable assignment: only keep lower bounds if RHS is a scalar
The current code was wrongly keeping the lower bounds of unallocated
arrays if the RHS was not a whole array variable. This is not what is
required 10.2.1.3 that states the bounds are kept only if the RHS
is a scalar (On top of being wrong, this was buggy if the lower
bounds had never been set before, which is possible with array RHS).