Fix allocatable assignment with vector subscript in RHS
The shape computation of allocatable assignment diverged from normal
array assignment shape deduction. Merge that back.
This revealed that the destShape set up in genarr(exv, components) was
wrong in case the `shape` was a shiftOp and that it ignored the slice.
Use genShape(arrayOperand) to fix that. This actually ddresses at least
part of the TODO(expected vector to have an extent) errors too outside
of allocatable assignment.

Also, when looking at the code that computes array shape in lowering, I
noticed there was some dead code/opportunities to simplify the code
by using `GetConstantExtents`. So I did that too.

The lit test fallouts are of two kinds:
- some convert from i64 -> index were removed due to the usage of
  `GetConstantExtents`.
- some slice extent computation were introduced. The comes from the fact that
  destShape is now computed in genarr() in cases were the shape was previously
  actually left empty and later computed by other means (e.g. another
  operand). This slice extent compuation is usually foldable by MLIR
  given it involves c1 operands.